// angular-biometric-designer/src/app/shared/models/system/agent.dto.ts
export class AgentDto {
    id: string = '';
    id_trans: string = '';
    agent_code: string = '';
    agent_name: string = '';
    branch_code: string = '';
    status: string = '';
    auth_stat: string = '';
    address: string = '';
    created_by: string = '';
    created_date: string = '';
    last_modified_by: string = '';
    last_modified_date: string = '';
    approved_by: string = '';
    approved_date: string = '';
  }